Added extra data to egg blocks

+++ src/main/java/com/barlinc/unusual_prehistory/blocks/WaterEggBlock.java
@@ -1,24 +1,30 @@
 package com.barlinc.unusual_prehistory.blocks;
 
+import com.barlinc.unusual_prehistory.blocks.blockentity.ExtraDataBlockEntity;
+import com.barlinc.unusual_prehistory.blocks.blockentity.FrogSpawnBlockEntity;
 import com.barlinc.unusual_prehistory.entity.base.PrehistoricAquaticMob;
 import com.barlinc.unusual_prehistory.entity.base.PrehistoricMob;
+import net.minecraft.advancements.CriteriaTriggers;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
+import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.tags.FluidTags;
 import net.minecraft.util.RandomSource;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.Mob;
-import net.minecraft.world.entity.MobSpawnType;
+import net.minecraft.world.entity.*;
 import net.minecraft.world.entity.animal.Animal;
-import net.minecraft.world.level.block.FrogspawnBlock;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.event.ForgeEventFactory;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
+import java.util.UUID;
 import java.util.function.Supplier;
 
-public class WaterEggBlock extends FrogspawnBlock {
+public class WaterEggBlock extends FrogSpawnBlockEntity {
 
     private final Supplier<EntityType<?>> hatchedEntity;
     private final int babyCount;
@@ -34,6 +40,9 @@ public class WaterEggBlock extends FrogspawnBlock {
         if (!this.canSurvive(state, level, pos)) {
             level.destroyBlock(pos, false);
         } else if (level.getFluidState(pos.below()).is(FluidTags.WATER)) {
+            BlockEntity blockEntity = level.getBlockEntity(pos);
+            if(!(blockEntity instanceof ExtraDataBlockEntity dataBlockEntity)) return;
+            UUID placer = dataBlockEntity.getOwner();
             level.destroyBlock(pos, false);
             for (int j = 0; j < babyCount; j++) {
                 Entity entity = hatchedEntity.get().create(level);
@@ -50,10 +59,32 @@ public class WaterEggBlock extends FrogspawnBlock {
                         prehistoricMob.setVariant(random.nextInt(prehistoricMob.getVariantCount()));
                     }
                     int k = random.nextInt(1, 361);
+                    if(placer != null) {
+                        Player player = level.getPlayerByUUID(placer);
+                        if(player instanceof ServerPlayer serverPlayer) {
+                            CriteriaTriggers.SUMMONED_ENTITY.trigger(serverPlayer, entity);
+                        }
+                    }
+
                     entity.moveTo(pos.getX(), (double) pos.getY() - 0.5D, pos.getZ(), (float) k, 0.0F);
                     ForgeEventFactory.onFinalizeSpawn(mob, level,level.getCurrentDifficultyAt(pos), MobSpawnType.NATURAL, null, null);
                 }
             }
         }
     }
+
+    @Override
+    public BlockEntity newBlockEntity(BlockPos pos, BlockState state) {
+        return new ExtraDataBlockEntity(pos, state);
+    }
+    @Override
+    public void setPlacedBy(Level pLevel, BlockPos pPos, BlockState pState, @Nullable LivingEntity pPlacer, ItemStack pStack) {
+        if (!pLevel.isClientSide && pPlacer instanceof Player player) {
+            BlockEntity be = pLevel.getBlockEntity(pPos);
+            if (be instanceof ExtraDataBlockEntity owned) {
+                owned.setOwner(player.getUUID());
+            }
+        }
+    }
+
 }
